# Shop_web_API/Shop_web_API/Models/DTOs/ProductDTO.cs
namespace Shop_web_API.Models.DTOs
{
    public class ProductDTO
    {
        public int Id { get; set; }
        public string Name { get; set; }
        public string Description { get; set; }
        public decimal Price { get; set; }
        public int Quantity { get; set; }
        public bool IsVisible { get; set; }

    }
}