# Shop_web_API/Shop_web_API/Models/OrderProduct.cs
namespace Shop_web_API.Models
{
    public class OrderProduct
    {
            public int OrderId { get; set; }
            public Order Order { get; set; }
            public int ProductId { get; set; }
            public Product Product { get; set; }
            public int Quantity { get; set; }
            public decimal TotalePrice { get; set; }
    }
}