# Shop_web_API/Shop_web_API/Models/Product.cs
namespace Shop_web_API.Models
{
        public class Product
        {
            public int Id { get; set; }
            public string Name { get; set; }
            public string Description { get; set; }
            public decimal Price { get; set; }
            public int Quantity { get; set; }
            public bool IsVisible { get; set; }
        
            public ICollection<OrderProduct> OrderProducts { get; set; }
        }
}