# Shop_web_API/Shop_web_API/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Shop Web API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Products": {
      "get": {
        "tags": [
          "Products"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductDTO"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProductDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        }
      }
    },
...
    "/api/Products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ProductDTO"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProductDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "ApplicationUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "normalizedUserName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "normalizedEmail": {
          "type": "string"
        },
        "emailConfirmed": {
          "type": "boolean"
        },
        "passwordHash": {
          "type": "string"
        },
        "securityStamp": {
          "type": "string"
        },
        "concurrencyStamp": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "phoneNumberConfirmed": {
          "type": "boolean"
        },
        "twoFactorEnabled": {
          "type": "boolean"
        },
        "lockoutEnd": {
          "format": "date-time",
          "type": "string"
        },
        "lockoutEnabled": {
          "type": "boolean"
        },
        "accessFailedCount": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
    "Order": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/ApplicationUser"
        },
        "orderDate": {
          "format": "date-time",
          "type": "string"
        },
        "orderProducts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderProduct"
          }
        }
      },
      "additionalProperties": false
    },
    "OrderProduct": {
      "type": "object",
      "properties": {
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "order": {
          "$ref": "#/definitions/Order"
        },
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "product": {
          "$ref": "#/definitions/Product"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "totalePrice": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
...
    "Product": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "isVisible": {
          "type": "boolean"
        },
        "orderProducts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderProduct"
          }
        }
      },
      "additionalProperties": false
    },
    "ProductDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "isVisible": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter 'Bearer' followed by a space and your token."
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}